passing urls to form
