+++ src/containers/AssetNewLoader.js
@@ -1,22 +1,14 @@
 /* eslint-disable no-console */
 import { connect } from 'react-redux'
 import { push } from 'connected-react-router'
+import { change, reset } from 'redux-form'
 import AssetNewForm from '../components/asset/AssetNew'
 import {
-    getActiveAccount,
-    putAsset,
-    handleBlobChosen
+    putAsset
 } from '../actions/index'
-import newAssetClearValues from '../constants'
 
 export default connect(
     state => ({
-        initialValues: {
-            ...state.newAsset,
-            selectBlobs: state.cloudStorage.blobs
-        },
-        activeAccount: getActiveAccount(state),
-        enableReinitialize: true,
         blobs: Object.values(state.cloudStorage.blobs).map((fileObject, index) => ({
             id: index,
             container: fileObject.container,
@@ -29,11 +21,11 @@ export default connect(
         onSubmit: values => {
             console.log('form values: ', values)
             dispatch(putAsset(values))
+            dispatch(reset('newAsset'))
             dispatch(push('/'))
-            dispatch({ type: 'GET_LINKS', links: '', ...newAssetClearValues })
         },
-        urlGetter: (blobsList) => {
-            dispatch(handleBlobChosen(blobsList))
+        linkSetter: (newValue) => {
+            dispatch(change('newAsset', 'contentUrls', newValue))
         }
     })
 )(AssetNewForm)
